requirements.txt
streamlit
python-dotenv
PyPDF2
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.1.0
faiss-cpu
streamlit-extras
sentence-transformers
torch
transformers
pydantic<2.0.0
groq
langchain-groq>=0.1.0